ftplugin: Improve dist tag replacement
If a spec file has 'Release: 1' the GetRelVer() function failed because
it stripped the entire release string.  The intent appears to be
removing dist tags which are often appended (e.g.: 1.fc25).

This is less than ideal for 2 reasons: 1) Not all spec files use a dist
tag; and 2) Many spec files which use a dist tag append values after the
dist tag (e.g.: 1.fc25.1).  The first case causes a traceback when
calling the function:

    Error detected while processing function <SNR>29_SpecChangelog..<SNR>29_GetRelVer:
    line   22:
    Traceback (most recent call last):
      File "<string>", line 19, in <module>
    vim.error: Vim(let):E15: Invalid expression:

The latter case would silently strip the .1 from the release, leaving
the changelog version string as 1.fc25.  This case is also the
documented way to perform a release bump on an older Fedora branch:
https://fedoraproject.org/wiki/Packaging:NamingGuidelines#Minor_release_bumps_for_old_branches

Use the expanded %dist macro for replacing the dist tag to avoid such
problems.

Signed-off-by: Todd Zullinger <tmz@pobox.com>